requirements.txt
django
gunicorn
djangorestframework